# docs/swagger.yaml
basePath: /
definitions:
...
  admin.createResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  admin.deleteResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  admin.listData:
    properties:
      created_at:
        description: 创建时间
        type: string
      created_user:
        description: 创建人
        type: string
      hashid:
        description: hashid
        type: string
      id:
        description: ID
        type: integer
      is_online:
        description: 是否在线 1:是 -1:否
        type: integer
      is_used:
        description: 是否启用 1:是 -1:否
        type: integer
      mobile:
        description: 手机号
        type: string
      nickname:
        description: 昵称
        type: string
      updated_at:
        description: 更新时间
        type: string
      updated_user:
        description: 更新人
        type: string
      username:
        description: 用户名
        type: string
    type: object
  admin.listResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/admin.listData'
        type: array
      pagination:
        properties:
          current_page:
            type: integer
          per_page_count:
            type: integer
          total:
            type: integer
        type: object
    type: object
...
  admin.modifyPasswordResponse:
    properties:
      username:
        description: 用户账号
        type: string
    type: object
...
  admin.offlineResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  authorized.createAPIResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  authorized.listAPIData:
    properties:
      api:
        description: 调用方对接人
        type: string
      business_key:
        description: 调用方key
        type: string
      hash_id:
        description: hashID
        type: string
      method:
        description: 调用方secret
        type: string
    type: object
  authorized.listAPIResponse:
    properties:
      business_key:
        description: 调用方key
        type: string
      list:
        items:
          $ref: '#/definitions/authorized.listAPIData'
        type: array
    type: object
...
  code.Failure:
    properties:
      code:
        description: 业务码
        type: integer
      data:
        description: 业务返回数据
      message:
        description: 描述信息
        type: string
    type: object
  config.emailResponse:
    properties:
      email:
        description: 邮箱地址
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/ChangSZ/gin-boilerplate/blob/master/LICENSE
  title: swagger 接口文档
  version: "2.0"
paths:
  /api/admin:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 管理员列表
      parameters:
      - default: 1
        description: 第几页
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: 每页显示条数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 昵称
        in: query
        name: nickname
        type: string
      - description: 手机号
        in: query
        name: mobile
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.listResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 管理员列表
      tags:
      - API.admin
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 新增管理员
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 昵称
        in: formData
        name: nickname
        required: true
        type: string
      - description: 手机号
        in: formData
        name: mobile
        required: true
        type: string
      - description: MD5后的密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.createResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 新增管理员
      tags:
      - API.admin
  /api/admin/{id}:
    delete:
      consumes:
      - application/json
      description: 删除管理员
      parameters:
      - description: hashId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.deleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 删除管理员
      tags:
      - API.admin
...
  /api/admin/modify_password:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: 修改密码
      parameters:
      - description: 旧密码
        in: formData
        name: old_password
        required: true
        type: string
      - description: 新密码
        in: formData
        name: new_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.modifyPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 修改密码
      tags:
      - API.admin
...
  /api/admin/offline:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: 下线管理员
      parameters:
      - description: Hashid
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.offlineResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 下线管理员
      tags:
      - API.admin
...
  /api/authorized_api:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 调用方接口地址列表
      parameters:
      - description: hashID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authorized.listAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 调用方接口地址列表
      tags:
      - API.authorized
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 授权调用方接口地址
      parameters:
      - description: HashID
        in: formData
        name: id
        required: true
        type: string
      - description: 请求方法
        in: formData
        name: method
        required: true
        type: string
      - description: 请求地址
        in: formData
        name: api
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authorized.createAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 授权调用方接口地址
      tags:
      - API.authorized
...
  /api/config/email:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: 修改邮件配置
      parameters:
      - description: 邮箱服务器
        in: formData
        name: host
        required: true
        type: string
      - description: 端口
        in: formData
        name: port
        required: true
        type: string
      - description: 发件人邮箱
        in: formData
        name: user
        required: true
        type: string
      - description: 发件人密码
        in: formData
        name: pass
        required: true
        type: string
      - description: 收件人邮箱地址，多个用,分割
        in: formData
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.emailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 修改邮件配置
      tags:
      - API.config
...
securityDefinitions:
  LoginToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"